+++ air/src/execution_step/mod.rs
@@ -29,7 +29,7 @@ const TETRAPLET_IDX_CORRECT: &str = "selects always return a correct index insid
 pub use errors::CatchableError;
 pub use errors::ExecutionError;
 pub use errors::UncatchableError;
-pub use execution_context::LastErrorObjectError;
+pub use execution_context::ErrorObjectError;
 pub use lambda_applier::LambdaError;
 
 pub mod errors_prelude {
@@ -51,7 +51,7 @@ pub(super) use value_types::ScalarRef;
 pub(super) use value_types::ServiceResultAggregate;
 pub(super) use value_types::Stream;
 pub(super) use value_types::ValueAggregate;
-pub(self) use value_types::STREAM_MAX_SIZE;
+use value_types::STREAM_MAX_SIZE;
 
 pub(crate) use air_trace_handler::TraceHandler;