+++ air/tests/test_module/instructions/fail.rs
@@ -15,6 +15,9 @@
  */
 
 use air::CatchableError;
+use air::ErrorObjectError;
+use air::ExecutionError;
+use air_test_framework::AirScriptExecutor;
 use air_test_utils::prelude::*;
 
 #[test]
@@ -28,6 +31,33 @@ fn fail_with_last_error() {
             (xor
                 (call "{local_peer_id}" ("service_id_1" "local_fn_name") [] result_1)
                 (fail %last_error%)
+            )"#
+    );
+
+    let result = call_vm!(vm, <_>::default(), script, "", "");
+
+    let expected_error = CatchableError::UserError {
+        error: rc!(json!({
+            "error_code": 10000i64,
+            "instruction": r#"call "local_peer_id" ("service_id_1" "local_fn_name") [] result_1"#,
+            "message": r#"Local service error, ret_code is 1, error message is '"failed result from fallible_call_service"'"#,
+            "peer_id": "local_peer_id",
+        })),
+    };
+    assert!(check_error(&result, expected_error));
+}
+
+#[test]
+fn fail_with_error() {
+    let local_peer_id = "local_peer_id";
+    let fallible_service_id = "service_id_1";
+    let mut vm = create_avm(fallible_call_service(fallible_service_id), local_peer_id);
+
+    let script = format!(
+        r#"
+            (xor
+                (call "{local_peer_id}" ("service_id_1" "local_fn_name") [] result_1)
+                (fail :error:)
             )"#
     );
 
@@ -98,6 +128,34 @@ fn fail_with_last_error_tetraplets() {
 }
 
 #[test]
+fn fail_with_error_tetraplets() {
+    let local_peer_id = "local_peer_id";
+    let fallible_service_id = "service_id_1";
+    let (host_closure, tetraplet_anchor) = tetraplet_host_function(fallible_call_service(fallible_service_id));
+    let mut vm = create_avm(host_closure, local_peer_id);
+
+    let local_fn_name = "local_fn_name";
+    let script = format!(
+        r#"
+        (xor
+            (xor
+                (call "{local_peer_id}" ("{fallible_service_id}" "{local_fn_name}") [] result_1)
+                (fail :error:)
+            )
+            (call "{local_peer_id}" ("" "") [%last_error%])
+        )
+          "#
+    );
+
+    let test_params = TestRunParameters::from_init_peer_id(local_peer_id);
+    let _ = checked_call_vm!(vm, test_params, script, "", "");
+    assert_eq!(
+        tetraplet_anchor.borrow()[0][0],
+        SecurityTetraplet::new(local_peer_id, fallible_service_id, local_fn_name, "")
+    );
+}
+
+#[test]
 fn fail_with_literals_tetraplets() {
     let local_peer_id = "local_peer_id";
     let (host_closure, tetraplet_anchor) = tetraplet_host_function(echo_call_service());
@@ -153,4 +211,59 @@ fn fail_with_canon_stream() {
         })),
     };
     assert!(check_error(&result, expected_error));
+}
+
+fn fail_to_fail_with_unsupported_errorcode(script: &str) {
+    let local_peer_id = "local_peer_id";
+    let script = script.to_string();
+
+    let executor = AirScriptExecutor::from_annotated(TestRunParameters::from_init_peer_id(local_peer_id), &script)
+        .expect("invalid test AIR script");
+    let results = executor.execute_all(local_peer_id).unwrap();
+
+    let expected_error = ExecutionError::Catchable(rc!(CatchableError::InvalidErrorObjectError(
+        ErrorObjectError::ErrorCodeMustBeNonZero
+    )));
+    assert!(check_error(&results.last().unwrap(), expected_error));
+}
+
+#[test]
+fn fail_to_fail_with_unsupported_errorcode_in_scalar() {
+    let script = r#"
+        (seq
+            (call "local_peer_id" ("m" "f1") [] scalar) ; ok = {"error_code": 0, "message": "some message"}
+            (fail scalar)
+        )
+    "#;
+    fail_to_fail_with_unsupported_errorcode(script);
+}
+
+#[test]
+fn fail_to_fail_with_unsupported_errorcode_in_scalar_wl() {
+    let script = r#"
+        (seq
+            (call "local_peer_id" ("m" "f1") [] scalar) ; ok = {"key": {"error_code": 0, "message": "some message"} }
+            (fail scalar.$.key)
+        )
+    "#;
+    fail_to_fail_with_unsupported_errorcode(script);
+}
+
+#[test]
+fn fail_to_fail_with_unsupported_errorcode_in_canon() {
+    let script = r#"
+        (seq
+            (call "local_peer_id" ("m" "f1") [] scalar) ; ok = [{"error_code": 0, "message": "some message"}]
+            (fail scalar.$.[0])
+        )
+    "#;
+    fail_to_fail_with_unsupported_errorcode(script);
+}
+
+#[test]
+fn fail_to_fail_with_unsupported_errorcode_in_error() {
+    let script = r#"
+        (fail :error:)
+    "#;
+    fail_to_fail_with_unsupported_errorcode(script);
 }