+++ crates/air-lib/execution-info-collector/CHANGELOG.md
@@ -16,6 +16,10 @@
   * dependencies
     * aquavm-air-parser bumped from 0.8.2 to 0.9.0
 
+* The following workspace dependencies were updated
+  * dependencies
+    * aquavm-air-parser bumped from 0.9.0 to 0.10.0
+
 ## [0.7.5](https://github.com/fluencelabs/aquavm/compare/air-execution-info-collector-v0.7.4...air-execution-info-collector-v0.7.5) (2023-06-22)