+++ crates/air-lib/polyplets/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1](https://github.com/fluencelabs/aquavm/compare/polyplets-v0.5.0...polyplets-v0.5.1) (2023-09-21)
+
+
+### Bug Fixes
+
+* **deps:** update rust crate marine-rs-sdk to 0.10.0 ([#640](https://github.com/fluencelabs/aquavm/issues/640)) ([b713e44](https://github.com/fluencelabs/aquavm/commit/b713e447fca38e0877a6c0e56bf91880f02bf9e4))
+
 ## [0.5.0](https://github.com/fluencelabs/aquavm/compare/polyplets-v0.4.0...polyplets-v0.5.0) (2023-08-17)